지금까지 선언한 type을 하나의 파일로 옮김
