+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { NavigationMenu } from "@/components/ui/navigation-menu";
 import TopBar from "@/ui/TopBar";
 
 const geistSans = Geist({
