+++ src/app/reports/page.tsx
@@ -1,8 +1,6 @@
 "use client";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { ImportStatReport } from "@/ui/ImportStatReport";
-import { StatReport } from "@/types/statReport";
-import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { PlusCircleIcon } from "lucide-react";
 import Link from "next/dist/client/link";
