--- src/types/matchRecord.ts
@@ -1,8 +1,0 @@
-import { ProfileRecord } from "@/types/profile";
-
-export type StatValue = Record<string, number>;
-export type MatchRecord = {
-  name: string;
-  enterDate: Date;
-  profileRecords: ProfileRecord[];
-};
