+++ src/types/statReport.ts
@@ -1,9 +1,15 @@
-import { StatDefinition } from "@/types/profile";
-import { MatchRecord } from "@/types/matchRecord";
-
-export type ProfileDefinition = {
+export type StatDefinition = { value: string };
+export type ProfileDefinition = { name: string; description?: string };
+export type StatValue = Record<string, number>;
+export type ProfileRecord = {
   name: string;
-  description?: string;
+  stats: StatValue;
+  count: number;
+};
+export type MatchRecord = {
+  name: string;
+  enterDate: Date;
+  profileRecords: ProfileRecord[];
 };
 export type StatReport = {
   name: string;
