+++ src/ui/ProfileDefinitionInput.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { TrashIcon, CirclePlusIcon } from "lucide-react";
-import { useFieldArray, useForm, useFormContext } from "react-hook-form";
+import { useFieldArray, useFormContext } from "react-hook-form";
 
 import { Button } from "@/components/ui/button";
 import { FormControl, FormField, FormItem, FormMessage } from "@/components/ui/form";
