+++ src/ui/ProfileInput.tsx
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button";
 import { FormControl, FormField, FormItem, FormMessage } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import React from "react";
-import { StatDefinition } from "@/types/profile";
+import { StatDefinition } from "@/types/statReport";
 
 export default function ProfileInput() {
   const { control, register, watch } = useFormContext();
